requirements.txt
# ONT Ecosystem - Core Requirements
# Install: pip install -r requirements.txt

pyyaml>=6.0
# Pin jsonschema<4.18 and referencing<0.30 for Python 3.10-3.12 compatibility
# (referencing>=0.30 uses TypeVar default= which requires Python 3.13+)
jsonschema>=4.0,<4.18
referencing<0.30
